+++ flatlaf-core/src/main/java/com/formdev/flatlaf/ui/FlatRootPaneUI.java
@@ -28,8 +28,6 @@ import java.awt.Insets;
 import java.awt.LayoutManager;
 import java.awt.LayoutManager2;
 import java.awt.Window;
-import java.awt.event.ComponentAdapter;
-import java.awt.event.ComponentEvent;
 import java.awt.event.ComponentListener;
 import java.beans.PropertyChangeEvent;
 import java.beans.PropertyChangeListener;
@@ -41,7 +39,6 @@ import javax.swing.JLayeredPane;
 import javax.swing.JMenuBar;
 import javax.swing.JRootPane;
 import javax.swing.LookAndFeel;
-import javax.swing.SwingUtilities;
 import javax.swing.UIManager;
 import javax.swing.border.Border;
 import javax.swing.plaf.BorderUIResource;
@@ -50,6 +47,7 @@ import javax.swing.plaf.RootPaneUI;
 import javax.swing.plaf.UIResource;
 import javax.swing.plaf.basic.BasicRootPaneUI;
 import com.formdev.flatlaf.FlatClientProperties;
+import com.formdev.flatlaf.FlatLaf;
 import com.formdev.flatlaf.util.HiDPIUtils;
 import com.formdev.flatlaf.util.SystemInfo;
 import com.formdev.flatlaf.util.UIScale;
@@ -87,9 +85,8 @@ public class FlatRootPaneUI
 
 	private Object nativeWindowBorderData;
 	private LayoutManager oldLayout;
-	private PropertyChangeListener ancestorListener;
-	private ComponentListener componentListener;
 	private ComponentListener macFullWindowContentListener;
+	private PropertyChangeListener macWindowBackgroundListener;
 
 	public static ComponentUI createUI( JComponent c ) {
 		return new FlatRootPaneUI();
@@ -107,6 +104,7 @@ public class FlatRootPaneUI
 			installBorder();
 
 		installNativeWindowBorder();
+		macInstallFullWindowContentSupport();
 	}
 
 	protected void installBorder() {
@@ -123,6 +121,7 @@ public class FlatRootPaneUI
 
 		uninstallNativeWindowBorder();
 		uninstallClientDecorations();
+		macUninstallFullWindowContentSupport();
 		rootPane = null;
 	}
 
@@ -156,6 +155,8 @@ public class FlatRootPaneUI
 			if( background == null || background instanceof UIResource )
 				parent.setBackground( UIManager.getColor( "control" ) );
 		}
+
+		macClearBackgroundForTranslucentWindow( c );
 	}
 
 	@Override
@@ -175,63 +176,20 @@ public class FlatRootPaneUI
 	protected void installListeners( JRootPane root ) {
 		super.installListeners( root );
 
-		if( SystemInfo.isJava_9_orLater ) {
-			// On HiDPI screens, where scaling is used, there may be white lines on the
-			// bottom and on the right side of the window when it is initially shown.
-			// This is very disturbing in dark themes, but hard to notice in light themes.
-			// Seems to be a rounding issue when Swing adds dirty region of window
-			// using RepaintManager.nativeAddDirtyRegion().
-			//
-			// Note: Not using a HierarchyListener here, which would be much easier,
-			// because this causes problems with mouse clicks in heavy-weight popups.
-			// Instead, add a listener to the root pane that waits until it is added
-			// to a window, then add a component listener to the window.
-			// See: https://github.com/JFormDesigner/FlatLaf/issues/371
-			ancestorListener = e -> {
-				Object oldValue = e.getOldValue();
-				Object newValue = e.getNewValue();
-				if( newValue instanceof Window ) {
-					if( componentListener == null ) {
-						componentListener = new ComponentAdapter() {
-							@Override
-							public void componentShown( ComponentEvent e ) {
-								// add whole root pane to dirty regions when window is initially shown
-								root.getParent().repaint( root.getX(), root.getY(), root.getWidth(), root.getHeight() );
-							}
-						};
-					}
-					((Window)newValue).addComponentListener( componentListener );
-				} else if( newValue == null && oldValue instanceof Window ) {
-					if( componentListener != null )
-						((Window)oldValue).removeComponentListener( componentListener );
-				}
-			};
-			root.addPropertyChangeListener( "ancestor", ancestorListener );
-		}
-
 		if( SystemInfo.isMacFullWindowContentSupported )
 			macFullWindowContentListener = FullWindowContentSupport.macInstallListeners( root );
+		macInstallWindowBackgroundListener( root );
 	}
 
 	@Override
 	protected void uninstallListeners( JRootPane root ) {
 		super.uninstallListeners( root );
 
-		if( SystemInfo.isJava_9_orLater ) {
-			if( componentListener != null ) {
-				Window window = SwingUtilities.windowForComponent( root );
-				if( window != null )
-					window.removeComponentListener( componentListener );
-				componentListener = null;
-			}
-			root.removePropertyChangeListener( "ancestor", ancestorListener );
-			ancestorListener = null;
-		}
-
 		if( SystemInfo.isMacFullWindowContentSupported ) {
 			FullWindowContentSupport.macUninstallListeners( root, macFullWindowContentListener );
 			macFullWindowContentListener = null;
 		}
+		macUninstallWindowBackgroundListener( root );
 	}
 
 	/** @since 1.1.2 */
@@ -311,17 +269,139 @@ public class FlatRootPaneUI
 
 	// layer title pane under frame content layer to allow placing menu bar over title pane
 	protected final static Integer TITLE_PANE_LAYER = JLayeredPane.FRAME_CONTENT_LAYER - 1;
+	private final static Integer TITLE_PANE_MOUSE_LAYER = JLayeredPane.FRAME_CONTENT_LAYER - 2;
+	private final static Integer WINDOW_TOP_BORDER_LAYER = Integer.MAX_VALUE;
+
+	// for fullWindowContent mode, layer title pane over frame content layer to allow placing title bar buttons over content
+	/** @since 3.4 */
+	protected final static Integer TITLE_PANE_FULL_WINDOW_CONTENT_LAYER = JLayeredPane.FRAME_CONTENT_LAYER + 1;
+
+	private Integer getLayerForTitlePane() {
+		return isFullWindowContent( rootPane ) ? TITLE_PANE_FULL_WINDOW_CONTENT_LAYER : TITLE_PANE_LAYER;
+	}
 
 	protected void setTitlePane( FlatTitlePane newTitlePane ) {
 		JLayeredPane layeredPane = rootPane.getLayeredPane();
 
-		if( titlePane != null )
+		if( titlePane != null ) {
 			layeredPane.remove( titlePane );
+			layeredPane.remove( titlePane.mouseLayer );
+			if( titlePane.windowTopBorderLayer != null )
+				layeredPane.remove( titlePane.windowTopBorderLayer );
+		}
 
-		if( newTitlePane != null )
-			layeredPane.add( newTitlePane, TITLE_PANE_LAYER );
+		if( newTitlePane != null ) {
+			layeredPane.add( newTitlePane, getLayerForTitlePane() );
+			layeredPane.add( newTitlePane.mouseLayer, TITLE_PANE_MOUSE_LAYER );
+			if( newTitlePane.windowTopBorderLayer != null && newTitlePane.isWindowTopBorderNeeded() && isFullWindowContent( rootPane ) )
+				layeredPane.add( newTitlePane.windowTopBorderLayer, WINDOW_TOP_BORDER_LAYER );
+		}
 
 		titlePane = newTitlePane;
+	}
+
+	private void macInstallFullWindowContentSupport() {
+		if( !SystemInfo.isMacOS )
+			return;
+
+		// set window buttons spacing
+		if( isMacButtonsSpacingSupported() && rootPane.isDisplayable() ) {
+			int buttonsSpacing = FlatNativeMacLibrary.BUTTONS_SPACING_DEFAULT;
+			String value = (String) rootPane.getClientProperty( FlatClientProperties.MACOS_WINDOW_BUTTONS_SPACING );
+			if( value != null ) {
+				switch( value ) {
+					case FlatClientProperties.MACOS_WINDOW_BUTTONS_SPACING_MEDIUM:
+						buttonsSpacing = FlatNativeMacLibrary.BUTTONS_SPACING_MEDIUM;
+						break;
+
+					case FlatClientProperties.MACOS_WINDOW_BUTTONS_SPACING_LARGE:
+						buttonsSpacing = FlatNativeMacLibrary.BUTTONS_SPACING_LARGE;
+						break;
+				}
+			}
+
+			FlatNativeMacLibrary.setWindowButtonsSpacing( getParentWindow( rootPane ), buttonsSpacing );
+		}
+
+		// update buttons bounds client property
+		FullWindowContentSupport.macUpdateFullWindowContentButtonsBoundsProperty( rootPane );
+	}
+
+	private void macUninstallFullWindowContentSupport() {
+		if( !SystemInfo.isMacOS )
+			return;
+
+		// do not uninstall when switching to another FlatLaf theme
+		if( UIManager.getLookAndFeel() instanceof FlatLaf )
+			return;
+
+		// reset window buttons spacing
+		if( isMacButtonsSpacingSupported() )
+			FlatNativeMacLibrary.setWindowButtonsSpacing( getParentWindow( rootPane ), FlatNativeMacLibrary.BUTTONS_SPACING_DEFAULT );
+
+		// remove buttons bounds client property
+		FullWindowContentSupport.macUninstallFullWindowContentButtonsBoundsProperty( rootPane );
+	}
+
+	private boolean isMacButtonsSpacingSupported() {
+		return SystemInfo.isMacOS && SystemInfo.isJava_17_orLater && FlatNativeMacLibrary.isLoaded();
+	}
+
+	private void macInstallWindowBackgroundListener( JRootPane c ) {
+		if( !SystemInfo.isMacOS )
+			return;
+
+		Window window = getParentWindow( c );
+		if( window != null && macWindowBackgroundListener == null ) {
+			macWindowBackgroundListener = e -> macClearBackgroundForTranslucentWindow( c );
+			window.addPropertyChangeListener( "background", macWindowBackgroundListener );
+		}
+	}
+
+	private void macUninstallWindowBackgroundListener( JRootPane c ) {
+		if( !SystemInfo.isMacOS )
+			return;
+
+		Window window = getParentWindow( c );
+		if( window != null && macWindowBackgroundListener != null ) {
+			window.removePropertyChangeListener( "background", macWindowBackgroundListener );
+			macWindowBackgroundListener = null;
+		}
+	}
+
+	/**
+	 * When setting window background to translucent color (alpha < 255),
+	 * Swing paints that window translucent on Windows and Linux, but not on macOS.
+	 * The reason for this is that FlatLaf sets the background color of the root pane,
+	 * and Swing behaves a bit differently on macOS than on other platforms in that case.
+	 * Other L&Fs do not set root pane background, which is {@code null} by default.
+	 * <p>
+	 * To fix this problem, set the root pane background to {@code null}
+	 * if windows uses a translucent background.
+	 */
+	private void macClearBackgroundForTranslucentWindow( JRootPane c ) {
+		if( !SystemInfo.isMacOS )
+			return;
+
+		Window window = getParentWindow( c );
+		if( window != null ) {
+			Color windowBackground = window.getBackground();
+			if( windowBackground != null &&
+				windowBackground.getAlpha() < 255 &&
+				c.getBackground() instanceof UIResource )
+			{
+				// clear root pane background
+				c.setBackground( null );
+			}
+		}
+	}
+
+	private Window getParentWindow( JRootPane c ) {
+		// not using SwingUtilities.windowForComponent() or SwingUtilities.getWindowAncestor()
+		// here because root panes may be nested and used anywhere (e.g. in JInternalFrame)
+		// but we're only interested in the "root" root pane, which is a direct child of the window
+		Container parent = c.getParent();
+		return (parent instanceof Window) ? (Window) parent : null;
 	}
 
 	@Override
@@ -368,6 +448,24 @@ public class FlatRootPaneUI
 					titlePane.titleBarColorsChanged();
 				break;
 
+			case FlatClientProperties.FULL_WINDOW_CONTENT:
+				if( titlePane != null ) {
+					rootPane.getLayeredPane().setLayer( titlePane, getLayerForTitlePane() );
+					if( titlePane.windowTopBorderLayer != null ) {
+						JLayeredPane layeredPane = rootPane.getLayeredPane();
+						if( titlePane.isWindowTopBorderNeeded() && isFullWindowContent( rootPane ) )
+							layeredPane.add( titlePane.windowTopBorderLayer, WINDOW_TOP_BORDER_LAYER );
+						else
+							layeredPane.remove( titlePane.windowTopBorderLayer );
+					}
+					titlePane.updateIcon();
+					titlePane.updateVisibility();
+					titlePane.updateFullWindowContentButtonsBoundsProperty();
+				}
+				FullWindowContentSupport.revalidatePlaceholders( rootPane );
+				rootPane.revalidate();
+				break;
+
 			case FlatClientProperties.FULL_WINDOW_CONTENT_BUTTONS_BOUNDS:
 				FullWindowContentSupport.revalidatePlaceholders( rootPane );
 				break;
@@ -382,56 +480,41 @@ public class FlatRootPaneUI
 				break;
 
 			case "ancestor":
+				if( e.getNewValue() instanceof Window )
+					macClearBackgroundForTranslucentWindow( rootPane );
+
+				macUninstallWindowBackgroundListener( rootPane );
+				macInstallWindowBackgroundListener( rootPane );
+
 				// FlatNativeMacLibrary.setWindowButtonsSpacing() and
 				// FullWindowContentSupport.macUpdateFullWindowContentButtonsBoundsProperty()
 				// require a native window, but setting the client properties
 				// "apple.awt.fullWindowContent" or FlatClientProperties.MACOS_WINDOW_BUTTONS_SPACING
 				// is usually done before the native window is created
 				// --> try again when native window is created
-				if( !SystemInfo.isMacOS || e.getNewValue() == null )
-					break;
-
-				// fall through
+				if( e.getNewValue() instanceof Window )
+					macInstallFullWindowContentSupport();
+				break;
 
 			case FlatClientProperties.MACOS_WINDOW_BUTTONS_SPACING:
-				if( SystemInfo.isMacOS ) {
-					// set window buttons spacing
-					if( SystemInfo.isJava_17_orLater && rootPane.isDisplayable() && FlatNativeMacLibrary.isLoaded() ) {
-						int buttonsSpacing = FlatNativeMacLibrary.BUTTONS_SPACING_DEFAULT;
-						String value = (String) rootPane.getClientProperty( FlatClientProperties.MACOS_WINDOW_BUTTONS_SPACING );
-						if( value != null ) {
-							switch( value ) {
-								case FlatClientProperties.MACOS_WINDOW_BUTTONS_SPACING_MEDIUM:
-									buttonsSpacing = FlatNativeMacLibrary.BUTTONS_SPACING_MEDIUM;
-									break;
-
-								case FlatClientProperties.MACOS_WINDOW_BUTTONS_SPACING_LARGE:
-									buttonsSpacing = FlatNativeMacLibrary.BUTTONS_SPACING_LARGE;
-									break;
-							}
-						}
-
-						Window window = SwingUtilities.windowForComponent( rootPane );
-						FlatNativeMacLibrary.setWindowButtonsSpacing( window, buttonsSpacing );
-					}
-
-					// update buttons bounds client property
-					FullWindowContentSupport.macUpdateFullWindowContentButtonsBoundsProperty( rootPane );
-				}
+				macInstallFullWindowContentSupport();
 				break;
 
 			case "apple.awt.fullWindowContent":
-				if( SystemInfo.isMacOS )
+				if( SystemInfo.isMacFullWindowContentSupported )
 					FullWindowContentSupport.macUpdateFullWindowContentButtonsBoundsProperty( rootPane );
 				break;
 		}
 	}
 
+	/** @since 3.4 */
+	protected static boolean isFullWindowContent( JRootPane rootPane ) {
+		return FlatClientProperties.clientPropertyBoolean( rootPane, FlatClientProperties.FULL_WINDOW_CONTENT, false );
+	}
+
 	protected static boolean isMenuBarEmbedded( JRootPane rootPane ) {
-		RootPaneUI ui = rootPane.getUI();
-		return ui instanceof FlatRootPaneUI &&
-			((FlatRootPaneUI)ui).titlePane != null &&
-			((FlatRootPaneUI)ui).titlePane.isMenuBarEmbedded();
+		FlatTitlePane titlePane = getTitlePane( rootPane );
+		return titlePane != null && titlePane.isMenuBarEmbedded();
 	}
 
 	/** @since 2.4 */
@@ -467,23 +550,21 @@ public class FlatRootPaneUI
 		private Dimension computeLayoutSize( Container parent, Function<Component, Dimension> getSizeFunc ) {
 			JRootPane rootPane = (JRootPane) parent;
 
-			Dimension titlePaneSize = (titlePane != null)
-				? getSizeFunc.apply( titlePane )
-				: new Dimension();
 			Dimension contentSize = (rootPane.getContentPane() != null)
 				? getSizeFunc.apply( rootPane.getContentPane() )
-				: rootPane.getSize();
+				: rootPane.getSize(); // same as in JRootPane.RootLayout.preferredLayoutSize()
 
 			int width = contentSize.width; // title pane width is not considered here
-			int height = titlePaneSize.height + contentSize.height;
+			int height = contentSize.height;
+			if( titlePane != null && !isFullWindowContent( rootPane ) )
+				height += getSizeFunc.apply( titlePane ).height;
 			if( titlePane == null || !titlePane.isMenuBarEmbedded() ) {
 				JMenuBar menuBar = rootPane.getJMenuBar();
-				Dimension menuBarSize = (menuBar != null && menuBar.isVisible())
-					? getSizeFunc.apply( menuBar )
-					: new Dimension();
-
-				width = Math.max( width, menuBarSize.width );
-				height += menuBarSize.height;
+				if( menuBar != null && menuBar.isVisible() ) {
+					Dimension menuBarSize = getSizeFunc.apply( menuBar );
+					width = Math.max( width, menuBarSize.width );
+					height += menuBarSize.height;
+				}
 			}
 
 			Insets insets = rootPane.getInsets();
@@ -508,12 +589,29 @@ public class FlatRootPaneUI
 			if( rootPane.getLayeredPane() != null )
 				rootPane.getLayeredPane().setBounds( x, y, width, height );
 
-			// title pane
+			// title pane (is a child of layered pane)
 			int nextY = 0;
 			if( titlePane != null ) {
 				int prefHeight = !isFullScreen ? titlePane.getPreferredSize().height : 0;
-				titlePane.setBounds( 0, 0, width, prefHeight );
-				nextY += prefHeight;
+				boolean isFullWindowContent = isFullWindowContent( rootPane );
+				if( isFullWindowContent && !UIManager.getBoolean( FlatTitlePane.KEY_DEBUG_SHOW_RECTANGLES ) ) {
+					// place title bar into top-right corner
+					int tw = Math.min( titlePane.getPreferredSize().width, width );
+					int tx = titlePane.getComponentOrientation().isLeftToRight() ? width - tw : 0;
+					titlePane.setBounds( tx, 0, tw, prefHeight );
+				} else
+					titlePane.setBounds( 0, 0, width, prefHeight );
+
+				titlePane.mouseLayer.setBounds( 0, 0, width, prefHeight );
+				if( titlePane.windowTopBorderLayer != null ) {
+					boolean show = isFullWindowContent && !titlePane.isWindowMaximized() && !isFullScreen;
+					if( show )
+						titlePane.windowTopBorderLayer.setBounds( 0, 0, width, 1 );
+					titlePane.windowTopBorderLayer.setVisible( show );
+				}
+
+				if( !isFullWindowContent )
+					nextY += prefHeight;
 			}
 
 			// glass pane
@@ -524,7 +622,7 @@ public class FlatRootPaneUI
 				rootPane.getGlassPane().setBounds( x, y + offset, width, height - offset );
 			}
 
-			// menu bar
+			// menu bar (is a child of layered pane)
 			JMenuBar menuBar = rootPane.getJMenuBar();
 			if( menuBar != null && menuBar.isVisible() ) {
 				boolean embedded = !isFullScreen && titlePane != null && titlePane.isMenuBarEmbedded();
@@ -532,13 +630,23 @@ public class FlatRootPaneUI
 					titlePane.validate();
 					menuBar.setBounds( titlePane.getMenuBarBounds() );
 				} else {
+					int mx = 0;
+					int mw = width;
+					if( titlePane != null && isFullWindowContent( rootPane ) ) {
+						// make menu bar width smaller to avoid that it overlaps title bar buttons
+						int tw = Math.min( titlePane.getPreferredSize().width, width );
+						mw -= tw;
+						if( !titlePane.getComponentOrientation().isLeftToRight() )
+							mx = tw;
+					}
+
 					Dimension prefSize = menuBar.getPreferredSize();
-					menuBar.setBounds( 0, nextY, width, prefSize.height );
+					menuBar.setBounds( mx, nextY, mw, prefSize.height );
 					nextY += prefSize.height;
 				}
 			}
 
-			// content pane
+			// content pane (is a child of layered pane)
 			Container contentPane = rootPane.getContentPane();
 			if( contentPane != null )
 				contentPane.setBounds( 0, nextY, width, Math.max( height - nextY, 0 ) );